requirements.txt
backports.csv==1.0.7
certifi==2019.6.16
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1
django-cors-headers==3.0.2
django-import-export==1.2.0
djangorestframework==3.9.3
et-xmlfile==1.0.1
jdcal==1.4.1
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.2
pytz==2019.1
PyYAML==5.1.1
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
